feat: implement complete multipart upload functionality and enhance media repository methods
