+++ backend/src/FileService/FileService.Contracts/CompleteMultipartUploadRequest.cs
@@ -1,0 +1,3 @@
+﻿namespace FileService.Contracts;
+
+public record CompleteMultipartUploadRequest(Guid MediaAssetId, string UploadId, IReadOnlyList<PartETagDto> PartETags);
