+++ backend/src/FileService/FileService.Core/DependencyInjection.cs
@@ -1,5 +1,6 @@
 using FileService.Core.Files;
 using FileService.Core.Files.FileKey;
+using FileService.Core.Multipart;
 using Microsoft.Extensions.Caching.Hybrid;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
@@ -11,6 +12,7 @@ public static class DependencyInjection
     public static IServiceCollection AddCore(this IServiceCollection services, IConfiguration configuration)
     {
         services.AddScoped<StartMultipartUploadHandler>();
+        services.AddScoped<CompleteMultipartUploadHandler>();
         services.AddSingleton<IFileKeyGenerator, FileKeyGenerator>();
 
         var redisConnectionString = configuration.GetConnectionString("Redis");
