+++ backend/src/FileService/FileService.Core/Multipart/IChunkSizeCalculator.cs
